Extract server-specific logic into a separate module
